Extra NULL safety
As the function is static and by inspection we can see it called with
a non-NULL value, the check is not valuable today. However, the check
is cheap (perhaps even optimized away) and if the function is later
used elsewhere, it will be nice to have the check in place.

Thanks to https://github.com/dgryski for the question!